requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
pymysql==1.1.0
cryptography==41.0.3
python-dotenv==1.0.0
pandas==2.1.0
alembic==1.12.0
pydantic==2.3.0
python-multipart==0.0.6
requests==2.31.0
